sqlite: cache column type/decl-type across queries
I spent a long time hunting for SQL query that could vary the returned
column type/decl-type conditional on parameters and could not find one.
So I believe this is safe.

	name        old time/op  new time/op  delta
	Persist-24  2.68µs ± 1%  2.45µs ± 1%  -8.39%  (p=0.000 n=20+20)

This benchmark is a bit of a tradeoff. On the one hand, there is only
one column type to cache so it undercounts the typical workload saved.
On the other hand, the query only returns one row, undercounting the
denominator.

There are plenty of real queries that return one row, so this has value.

Signed-off-by: David Crawshaw <crawshaw@tailscale.com>