sqlite, etc: add BusyTimeout
I am somewhat at a loss for why setting a busy timeout seems to be
required for standard concurrent operation of writers. According to the
documentation:

	https://sqlite.org/wal.html#sometimes_queries_return_sqlite_busy_in_wal_mode

The conditions under which a WAL database should block committing
transactions should be rare. None of the test cases I have written
trigger those conditions. So why do I have to set a busy timeout to
avoid almost all concurrent write transactions getting caught on
SQLITE_BUSY?

Anyway, it is a mystery that can be investigated slowly. The busy
timeout does not inhibit good throughput, it is easy to reach 50k write
transactions per second on a good SSD, with this busy timeout set.

Signed-off-by: David Crawshaw <crawshaw@tailscale.com>