sqlite: really close
sqlite3_close doesn't close the DB unless all the prepared statements
are finalized. Callers of the package can get this right for
non-persistent queries, but for the persistent ones we need to clear our
internal cache.

This doesn't have a test because the sql.Conn.Close method eats errors
from the driver. Easily manually testable by taking SQL-based tests and
running them with -count=1000, as you run out of fds.

Signed-off-by: David Crawshaw <crawshaw@tailscale.com>