sqlitestats: new package with Tracer web interface
Initial implementation is rudimentary, but useful if you lock up your DB
with interdependent write transactions.

Follow up work will include a "collector" that traces out all
transactions for a 1-2 second period on http request. That needs a bit
of care in the design, as we have enough read-only traffic that we need
to coalesce similar tx's (like the pprof view of Go stack traces.)