sqlitepool: new package, implements a pool
This is intended as a replacement for our internal sqlitedb package
pool and its TxWrapper.

It needs a little more machinery to be dropped in that's not covered in
this PR, in particular:
- helper functions that mimick ExecContext/QueryContext/QueryRowContext
  to make replacing the existing TxWrappers easy
- a little more Tracer plumbing, so we can get SQL query exec times
  out in the same format we are used to them now
- a few other missing counters

This PR implements the Tx/Rx object split we've discussed a few times.
It will take a bit of work to make good use of that internally. But I
figured it's worth getting in now.