sqlitepool: add transitionary query glue
This adds Exec/Query/QueryRow that function similarly to the core
methods of the database/sql package, to make transitioning from the
sqlite driver to sqlitepool easier.

The long-term goal is to replace uses of this with generated code
directly on top of the APIs that require fewer allocations to use. But
there are enough touch points that having these around will let us do
the transition in stages.