cgosqlite,sqlite: expose incremental API and add documentation
It is likely useful to report progress information and offer abort for
large databases, so expose this in the high level API in a somewhat safe
way.

A lock was not introduced into the driver directly at this time because
doing so is much more complex, particularly as Stmt.DBHandle would need
to recover that lock - essnetially requiring global state with further
synchronization.