cgosqlite: move WAL callback to its own file
As per the last comment of golang/go#48824, _GoStringLen and
_GoStringPtr are declared in the preamble unless //export is used. To
fix the problem, the exported WAL callback function needs to be moved to
a separate file.

This removes compiler errors under certain conditions. The most easy way
to reproduce this is by attempting to use dlv. For some reason, a normal
go build still succeeds but a dlv build does not.

Fixes #75

Signed-off-by: Adrian Dewhurst <sailor@sailorfrag.net>