cgosqlite: import SQLite amalgamation v3.42.0 shell.c
We are going to start storing the matching SQLite amalgamation shell.c
file in the repository, even though it's not compiled into the package,
so that users can build an sqlite3 binary that exactly matches the
SQLite they are using in their Go programs.

The file is imported as shell.c.disabled so it isn't picked up by cgo.

Signed-off-by: David Crawshaw <crawshaw@tailscale.com>