sqlite: cancel completed queries after cleanup, not before
In #85 we added an optional mechanism to forcibly interrupt a query that
exceeds its context's lifespan. Move the cancellation during query close until
after cleanup, so that the cancellation of a successful query does not race
with cleanup for setting the result code. (It is not a data race, the accesses
are properly synchronized, but we don't want the cancellation to win unless the
query is actually still running).