cgosqlite: do not attempt to reset a NULL statement
Presenting a query consisting of only comments results in a statement with a
NULL underlying cStmt. The SQL text is syntactically valid, but when attempting
to clean up after a query on such a statement, the indirection through NULL
triggers a panic.

Check during cleanup for this degenerate case, and don't try to reset the
statement in that case. Add a regression test for the original failure.

Signed-off-by: M. J. Fromberger <fromberger@tailscale.com>