cgosqlite: add more string interns, and optimize
This adds string interning to column names, table names and database
names, as well as optimizing the decltype interning code path to use
runtime.findnull rather than C.strlen that must cross cgo again.

We don't yet know if this is effective in practice, and should test this
before integrating.