sqlite: fix cancellation cleanup for ExecContext (#104)
I did not test the changes in #103 sufficiently, and missed a corner case in
ExecContext: We need to make sure the context cleanup happens prior to the
synchronization, not in a defer.

Add a test to actually trigger this case, and fix the cleanup so it happens
before return as intended. Add some missing documentation too.